Add unit test. Extract api logic to service
