+++ VoxelSniperCore/src/test/java/com/github/kevindagame/util/VersionCheckerTest.kt
@@ -1,9 +1,12 @@
 package com.github.kevindagame.util
 
 
+import com.github.kevindagame.service.GithubService
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
+import org.mockito.Mockito.mock
+import org.mockito.Mockito.`when`
 
 class VersionCheckerTest {
     @Test
@@ -27,6 +30,15 @@ class VersionCheckerTest {
         assertTrue(v1 != v2)
         assertTrue(v2 != v3)
         assertTrue(v3 != v4)
+    }
+
+    @Test
+    fun getLatestVersionInfo_invalid_version() {
+        val githubService = mock(GithubService::class.java)
+        `when`(githubService.getReleasesFromApi()).thenReturn(listOf(Release("", "v1.0.0", emptyList<Asset>(), false)))
+        val versionChecker = VersionChecker(githubService)
+        val latestVersionInfo = versionChecker.getLatestVersionInfo("v1.1.0-SNAPSHOT")
+        assertEquals(null, latestVersionInfo)
     }
 
 }
